quantizer: input length type changed from uint16_t to int32_t

+++ examples/example_noise_shaping.cpp
@@ -13,11 +13,11 @@ void setup() {
   quantizer.configure(noiseShaping, dither, factor);
   double frequ=4000;
   double samplesPerPeriod=44100/frequ;
-  uint16_t noSamples=(uint16_t)round(200*samplesPerPeriod);
+  int32_t noSamples=(int32_t)round(200*samplesPerPeriod);
   float samples[noSamples];
 
 
-  for (uint16_t i=0; i< noSamples; i++){
+  for (int32_t i=0; i< noSamples; i++){
     samples[i]=sin(i*frequ/fs*TWO_PI)*0.9;
   }
   int16_t outputSamples[noSamples];
@@ -27,7 +27,7 @@ void setup() {
   Serial.print("ellapsed time: ");
   Serial.println(t1-t0);
 
-  for (uint32_t i =0 ; i< noSamples; i++){
+  for (int32_t i =0 ; i< noSamples; i++){
       //Serial.print(outputSamples[2*i+1]);     
       Serial.print(outputSamples[i]);     
       Serial.println(",");
